Ignore style/script elements by default, allow opting for other elements too
script/style are typically not part of the content of a page, yet they can be quite large blocks of text that would consume memory unnecessarily.

We should skip both by default, and allow opting in for those as well as potentially other elements via settings.

This commit reworks the way we deal with settings so all overloads have a consistent and non-duplicated behavior.

Defaulting the HTML 4.01 transitional DTD also allows for much cleaner object model, so we opt in for that too.

Fixes #8